@@ -1,7 +1,7 @@
 {
   "name": "nastia",
   "version": "0.0.1",
-  "description": "",
+  "description": "Nastia legacy website",
   "main": "index.js",
   "scripts": {
     "clean": "rm -rf ./public/dist ./client/dist",
@@ -13,29 +13,33 @@
   "license": "ISC",
   "repository": "https://github.com/Lithium366/nastya-website.git",
   "dependencies": {
-    "compression": "^1.6.0",
-    "ejs": "^2.3.4",
-    "express": "^4.13.3",
+    "compression": "^1.7.4",
+    "ejs": "^2.6.2",
+    "express": "^4.17.1",
     "express-partials": "^0.3.0",
-    "history": "^1.17.0",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-ga": "^1.2.1",
-    "react-router": "^1.0.3"
+    "history": "^4.9.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-ga": "^2.6.0",
+    "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
-    "babel-plugin-transform-react-jsx": "^6.3.13",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "grunt": "^0.4.5",
-    "grunt-babel": "^6.0.0",
-    "grunt-browserify": "^4.0.1",
-    "grunt-concurrent": "^2.1.0",
-    "grunt-contrib-uglify": "^0.11.0",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-nodemon": "^0.4.1",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-transform-react-jsx": "^7.3.0",
+    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-es2015": "^7.0.0-beta.53",
+    "@babel/preset-react": "^7.0.0",
+    "grunt": "^1.0.4",
+    "grunt-babel": "^8.0.0",
+    "grunt-browserify": "^5.3.0",
+    "grunt-concurrent": "^3.0.0",
+    "grunt-contrib-uglify": "^4.0.1",
+    "grunt-contrib-watch": "^1.1.0",
+    "grunt-nodemon": "^0.4.2",
     "grunt-reactify": "0.0.2",
-    "grunt-sass": "^1.1.0",
-    "load-grunt-tasks": "^3.4.0"
+    "grunt-sass": "^3.0.2",
+    "load-grunt-tasks": "^5.0.0",
+    "node-sass": "^4.12.0",
+    "react-dom": "^16.8.6"
   }
 }
