@@ -1,27 +1,23 @@
 "use strict";
 
 import React from 'react';
-import { render } from 'react-dom'
-import { Router, Route, Link } from 'react-router';
-const createBrowserHistory = require('history/lib/createBrowserHistory');
+import ReactDOM from 'react-dom';
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
+import Index from './pages/index.jsx';
+import Resume from './pages/resume.jsx';
+import OtherWorks from './pages/other.jsx';
+
 const ga = require('react-ga');
-
-const Index = require('./pages/index.jsx');
-const Resume = require('./pages/resume.jsx');
-const OtherWorks = require('./pages/other.jsx');
-
 ga.initialize('UA-71640692-1');
 
 const logPageView = function () {
     ga.pageview(window.location.pathname);
 };
 
-render((
-    <Router history={createBrowserHistory()} onUpdate={logPageView}>
-        <Route path="/" component={Index} />
+ReactDOM.render((
+    <Router onUpdate={logPageView}>
+        <Route path="/" exact component={Index} />
         <Route path="/resume" component={Resume} />
-        <Route path="/portfolio" component={OtherWorks}>
-            <Route path="/portfolio/:section" component={OtherWorks} />
-        </Route>
+        <Route path="/portfolio/:section?" component={OtherWorks} />
     </Router>
 ), document.getElementById('mainContainer'));
